requirements.txt
django
djangorestframework
django-filter
python 3.6
